requirements-rpi.txt
dbus-python==1.2.12
googlemaps==4.0.0
Kivy-Garden==0.1.4
plyer==1.4.2
polyline==1.4.0
pulsectl==19.10.4
